ml bayes fix, model updated
