+++ src/aop_usage_example.py
@@ -7,4 +7,4 @@ ml = MachineLearning()
 aspect.apply(ml)
 print(ml.analyze("You are an idiot"))
 print(ml.analyze("Hello friend"))
-print(ml.analyze("This man is a sickening, mentally ill joke"))
+
